pyproject.toml
[project]
name = "bakalarka-gtfs-agent"
version = "0.3.0"
description = "MCP server a custom GTFS agent pre bakalarsku pracu"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.2.0",
    "openai-agents>=0.0.7",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9.0",
    "pre-commit>=4.0.0",
]

[project.scripts]
gtfs-mcp = "bakalarka_gtfs.mcp.server:main"
gtfs-api = "bakalarka_gtfs.api.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/bakalarka_gtfs"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# ---------------------------------------------------------------------------
# Ruff — linter + formatter
# ---------------------------------------------------------------------------

[tool.ruff]
target-version = "py311"
src = ["src", "tests"]
line-length = 120

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort (import ordering)
    "N",      # pep8-naming
    "UP",     # pyupgrade (modern Python syntax)
    "B",      # flake8-bugbear (common bugs)
    "A",      # flake8-builtins (shadowing builtins)
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking (TYPE_CHECKING blocks)
    "RUF",    # ruff-specific rules
]
ignore = [
    "E501",   # line too long — handled by formatter
    "B008",   # do not perform function calls in argument defaults (FastAPI Depends)
    "A003",   # class attribute shadowing builtin (e.g. id, type)
    "SIM108", # ternary — niekedy je if/else čitateľnejší
    "N802",   # function name should be lowercase — výnimky pre existujúce pomenovania
    "N806",   # variable should be lowercase — výnimky
    "RUF012", # mutable class attributes — dataclass fields sú OK
]

[tool.ruff.lint.isort]
known-first-party = ["bakalarka_gtfs"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
